refactor: update user entity imports and restructure OAuth2 related files
