+++ src/auth/auth.module.ts
@@ -10,10 +10,10 @@ import { TwoFactorController } from './two-factor.controller';
 import { JwtStrategy } from './jwt.strategy';
 import { JwtAuthGuard } from './jwt-auth.guard';
 import { PermissionsGuard } from './permissions.guard';
-import { User } from '../user/user.entity';
-import { Client } from '../client/client.entity';
-import { Token } from '../token/token.entity';
-import { AuthorizationCode } from '../authorization-code/authorization-code.entity';
+import { User } from './user.entity';
+import { Client } from '../oauth2/client.entity';
+import { Token } from '../oauth2/token.entity';
+import { AuthorizationCode } from '../oauth2/authorization-code.entity';
 import { LoggingModule } from '../logging/logging.module';
 import { UtilsModule } from '../utils/utils.module';
 import { JWT_CONSTANTS } from '../constants/auth.constants';
