+++ src/auth/auth.service.spec.ts
@@ -4,10 +4,10 @@ import { JwtService } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
 import { CACHE_MANAGER } from '@nestjs/cache-manager';
 import { AuthService } from './auth.service';
-import { User } from '../user/user.entity';
-import { Client } from '../client/client.entity';
-import { Token } from '../token/token.entity';
-import { AuthorizationCode } from '../authorization-code/authorization-code.entity';
+import { User } from './user.entity';
+import { Client } from '../oauth2/client.entity';
+import { Token } from '../oauth2/token.entity';
+import { AuthorizationCode } from '../oauth2/authorization-code.entity';
 import { TwoFactorService } from './two-factor.service';
 import { FileUploadService } from '../upload/file-upload.service';
 import { RecaptchaService } from '../utils/recaptcha.util';
