+++ src/auth/auth.service.ts
@@ -12,10 +12,10 @@ import { Repository } from 'typeorm';
 import { CACHE_MANAGER } from '@nestjs/cache-manager';
 import type { Cache } from 'cache-manager';
 import { JwtService } from '@nestjs/jwt';
-import { User } from '../user/user.entity';
-import { Client } from '../client/client.entity';
-import { Token } from '../token/token.entity';
-import { AuthorizationCode } from '../authorization-code/authorization-code.entity';
+import { User } from './user.entity';
+import { Client } from '../oauth2/client.entity';
+import { Token } from '../oauth2/token.entity';
+import { AuthorizationCode } from '../oauth2/authorization-code.entity';
 import { TokenDto } from './dto/response.dto';
 import * as bcrypt from 'bcrypt';
 import * as speakeasy from 'speakeasy';
