+++ src/auth/two-factor.service.ts
@@ -8,7 +8,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import * as speakeasy from 'speakeasy';
 import * as QRCode from 'qrcode';
-import { User } from '../user/user.entity';
+import { User } from './user.entity';
 import { TwoFactorResponseDto } from './dto/2fa/two-factor.dto';
 import {
   AUTH_ERROR_MESSAGES,
