+++ src/oauth2/oauth2.module.ts
@@ -18,11 +18,11 @@ import { CleanupSchedulerService } from './cleanup-scheduler.service';
 import { OAuth2BearerGuard } from './guards/oauth2-bearer.guard';
 import { OAuth2ScopeGuard } from './guards/oauth2-scope.guard';
 import { OAuth2Strategy } from './guards/oauth2.strategy';
-import { User } from '../user/user.entity';
-import { Client } from '../client/client.entity';
-import { AuthorizationCode } from '../authorization-code/authorization-code.entity';
-import { Token } from '../token/token.entity';
-import { Scope } from '../scope/scope.entity';
+import { User } from '../auth/user.entity';
+import { Client } from './client.entity';
+import { AuthorizationCode } from './authorization-code.entity';
+import { Token } from './token.entity';
+import { Scope } from './scope.entity';
 import { AppConfigService } from '../config/app-config.service';
 import { LoggingModule } from '../logging/logging.module';
 import { CacheConfigModule } from '../cache/cache-config.module';
