+++ src/oauth2/scope.entity.ts
