+++ src/oauth2/services/token-grant.service.ts
@@ -1,8 +1,8 @@
 import { Injectable, BadRequestException, Logger } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { User } from '../../user/user.entity';
-import { Client } from '../../client/client.entity';
+import { User } from '../../auth/user.entity';
+import { Client } from '../client.entity';
 import { AuthorizationCodeService } from '../authorization-code.service';
 import { TokenService } from '../token.service';
 import { ScopeService } from '../scope.service';
