+++ src/oauth2/token.service.ts
@@ -6,9 +6,9 @@ import { ConfigService } from '@nestjs/config';
 import { CACHE_MANAGER } from '@nestjs/cache-manager';
 import type { Cache } from 'cache-manager';
 import { AppConfigService } from '../config/app-config.service';
-import { Token } from '../token/token.entity';
-import { User } from '../user/user.entity';
-import { Client } from '../client/client.entity';
+import { Token } from './token.entity';
+import { User } from '../auth/user.entity';
+import { Client } from './client.entity';
 import { OAuth2JwtPayload } from '../types/oauth2.types';
 import { StructuredLogger } from '../logging/structured-logger.service';
 import { TOKEN_TYPES, JWT_TOKEN_EXPIRY } from '../constants/auth.constants';
